removed all core components folders refactored components to match export changed imports in the code
