+++ src/app/[lang]/library/create-library-content/create-post/page.tsx
@@ -1,7 +1,8 @@
 'use client';
-import Button from '@/components/core/button';
-import Input from '@/components/core/input/Input';
-import Link from '@/components/core/link';
+import { Button } from '@/components/core';
+
+import { Input } from '@/components/core';
+import { Link } from '@/components/core';
 import { PostType } from '@/types/libraryPageTypes/contentElements';
 import React from 'react';
 import { POST } from '../../../../api/library-content/route';
