+++ src/app/[lang]/library/posts/[id]/page.tsx
@@ -1,12 +1,12 @@
-import Input from '@/components/core/input/Input';
-import Link from '@/components/core/link';
+import { Input } from '@/components/core';
+import { Link } from '@/components/core';
 import React from 'react';
 
 const PostPage = () => {
     return (
         <div>
-                full post 
-          
+            full post
+
             <Link label='back to library' url='/library/library/posts' />
         </div>
     );
