+++ src/components/core/Menubar.tsx
@@ -1,11 +1,12 @@
-import React, { useState } from 'react';
+'use client';
+import { useState } from 'react';
 
 type Props = {
 	content: React.ReactNode;
 	trigger: React.ReactNode;
 };
 
-const Menubar = ({ content, trigger }: Props) => {
+export const Menubar = ({ content, trigger }: Props) => {
 	const [isOpen, setIsOpen] = useState(false);
 
 	const handleTriggerClick = () => {
@@ -34,9 +35,8 @@ const Menubar = ({ content, trigger }: Props) => {
 				{trigger}
 			</div>
 			<div
-				className={`absolute ${
-					isOpen ? 'block' : 'hidden'
-				} bg-light-surface-primary dark:bg-dark-surface-primary text-light-surface-onPrimary dark:text-dark-surface-onPrimary absolute top-9 mt-3 min-w-[120px] rounded border py-2`}
+				className={`absolute ${isOpen ? 'block' : 'hidden'
+					} bg-light-surface-primary dark:bg-dark-surface-primary text-light-surface-onPrimary dark:text-dark-surface-onPrimary absolute top-9 mt-3 min-w-[120px] rounded border py-2`}
 				onClick={handleContentClick}
 			>
 				{content}
@@ -44,5 +44,3 @@ const Menubar = ({ content, trigger }: Props) => {
 		</button>
 	);
 };
-
-export default Menubar;
