+++ src/components/core/Text.tsx
@@ -1,0 +1,13 @@
+type Props = {
+  children: React.ReactNode;
+  className?: string;
+};
+
+export default function Text({ children, className = '' }: Props) {
+  return (
+    <p
+      className={`text-light-surface-onPrimary dark:text-dark-surface-onPrimary font-inter ${className}`}>
+      {children}
+    </p>
+  );
+}
