--- src/components/core/card/index.ts
@@ -1,3 +1,0 @@
-import Card from './Card';
-
-export default Card;
