--- src/components/core/circleGradient/index.tsx
@@ -1,2 +1,0 @@
-import CircleGradient from './CircleGradient';
-export default CircleGradient;
