--- src/components/core/levelLink/index.ts
@@ -1,2 +1,0 @@
-import levelLink from "./LevelLink";
-export default levelLink;
