--- src/components/core/link/index.ts
@@ -1,2 +1,0 @@
-import Link from '../link/Link';
-export default Link;
