--- src/components/core/menubar/index.ts
@@ -1,3 +1,0 @@
-import Menubar from './Menubar';
-
-export default Menubar;
