+++ src/components/core/notesCircle/index.ts
@@ -1,0 +1,2 @@
+import NotesCircle from './NotesCircle';
+export default NotesCircle;
