--- src/components/core/progressBar/index.ts
@@ -1,2 +1,0 @@
-import ProgressBar from './progressBar';
-export default ProgressBar;
