--- src/components/core/statsCard/index.ts
@@ -1,2 +1,0 @@
-import StatsCard from "./StatsCard";
-export default StatsCard;
