+++ src/components/shared/libraryFilter/LibraryFilter.tsx
@@ -1,6 +1,6 @@
 import { FilterObject } from '@/types';
-import Button from '@/components/core/button';
-import Input from '@/components/core/input/Input';
+import { Button } from '@/components/core';
+import { Input } from '@/components/core';
 
 type Props = {
     filterButtons: FilterObject;
