+++ src/components/shared/menubar/Menubar.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import Card from '@core/card';
+import { Card } from '@/components/core';
 import { Props as MenubarItemProps } from './MenubarItem';
 import MenubarItem from './MenubarItem';
 import { useState } from 'react';
