+++ src/components/shared/navbar/mobileNav/MobileNav.tsx
@@ -1,5 +1,5 @@
 import { Locale } from '@/i18n.config';
-import Card from '@core/card';
+import { Card } from '@/components/core';
 import { CloseIcon } from '@icons';
 import { useState } from 'react';
 import MobileNavContent from './MobileNavContent';
